use geometry of local area to derive orientation of cart
This is based on the assumption that the cart is placed against a
wall/edge in the local area specified in the task.
We first find the edge in the local area closest to the pose of
the cart, then set the pose of the cart as close to the orientation
of the normal of the edge (in 90 degree offsets from the original
cart orientation).